walkrandomlogic > walkrandomlogic.cs
